pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aether-runner"
version = "0.3.5"
description = "Execute Luau tests on Roblox Cloud"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    {name = "WildLink Team"}
]
dependencies = [
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "watchdog>=3.0.0",
    "tomli>=2.0.0; python_version < '3.11'"
]

[project.scripts]
aether = "aether.cli:main"

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "mypy>=1.0.0",
    "pyinstaller>=6.0.0"
]
